+++ packages/ragbits/CHANGELOG.md
@@ -2,6 +2,39 @@
 
 ## Unreleased
 
+## 0.19.1 (2025-05-27)
+
+### Changed
+
+- ragbits-chat updated to version v0.19.1
+- ragbits-cli updated to version v0.19.1
+- ragbits-document-search updated to version v0.19.1
+- ragbits-evaluate updated to version v0.19.1
+- ragbits-guardrails updated to version v0.19.1
+- ragbits-core updated to version v0.19.1
+
+## 0.19.0 (2025-05-27)
+
+### Changed
+
+- ragbits-chat updated to version v0.19.0
+- ragbits-cli updated to version v0.19.0
+- ragbits-document-search updated to version v0.19.0
+- ragbits-evaluate updated to version v0.19.0
+- ragbits-guardrails updated to version v0.19.0
+- ragbits-core updated to version v0.19.0
+
+## 0.18.0 (2025-05-22)
+
+### Changed
+
+- ragbits-chat updated to version v0.18.0
+- ragbits-cli updated to version v0.18.0
+- ragbits-document-search updated to version v0.18.0
+- ragbits-evaluate updated to version v0.18.0
+- ragbits-guardrails updated to version v0.18.0
+- ragbits-core updated to version v0.18.0
+
 ## 0.17.1 (2025-05-09)
 
 ### Changed